Revert "Readme updates"
