+++ client/public/index.html
@@ -3,28 +3,44 @@
 
 <head>
   <meta charset="utf-8">
-  <link rel="shortcut icon" type="image/png" href="./favicon_blue.png">
-  <link rel="icon" type="image/png" href="./favicon_blue.png">
+  <link rel="shortcut icon" type="image/png" href="./favicon.png">
+  <link rel="icon" type="image/png" href="./favicon_allerg.jpg">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
   <meta name="theme-color" content="#000000">
+  <!--
+      manifest.json provides metadata used when your web app is added to the
+      homescreen on Android. See https://developers.google.com/web/fundamentals/engage-and-retain/web-app-manifest/
+    -->
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-  
-  <!--Bootstrap & Google Fonts-->
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy"
     crossorigin="anonymous"></script>
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
     crossorigin="anonymous">
+  <!-- Google Fonts -->
+  <!-- <link href="https://fonts.googleapis.com/css?family=Poppins:300,600" rel="stylesheet"> -->
   <link href="https://fonts.googleapis.com/css?family=Montserrat|Open+Sans" rel="stylesheet">
+   <!--Custom CSS--> 
+   <!-- <link rel="stylesheet" src="../src/assets/css/styles.css"> -->
 
-  <!-- Custom CSS -->
-  <link rel="stylesheet" src="../src/assets/css/style.css">
-  <link rel="stylesheet" href="https://unpkg.com/@coreui/icons/css/coreui-icons.min.css">
+   <!-- ERIN CSS -->
+   <link rel="stylesheet" src="../src/assets/css/style.css">
+   <link rel="stylesheet" href="https://unpkg.com/@coreui/icons/css/coreui-icons.min.css">
+  
+  <!-- <script src="../js/react.min.js"></script> -->
+    <!-- <script src="../js/react-dom.min.js"></script> -->
+    <!-- <script src="../js/browser.min.js"></script> -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/6.1.19/browser.min.js"></script>
+    <script crossorigin src="https://unpkg.com/react@16/umd/react.production.min.js"></script>
+    <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js"></script>
+  <!--
+      Notice the use of %PUBLIC_URL% in the tags above.
+      It will be replaced with the URL of the `public` folder during the build.
+      Only files inside the `public` folder can be referenced from the HTML.
 
-  <!--Babel and React-->
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/6.1.19/browser.min.js"></script>
-  <script crossorigin src="https://unpkg.com/react@16/umd/react.production.min.js"></script>
-  <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js"></script>
-
+      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
+      work correctly both with client-side routing and a non-root public URL.
+      Learn how to configure a non-root public URL by running `npm run build`.
+    -->
   <title>Allergenius</title>
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
     crossorigin="anonymous">
@@ -32,7 +48,7 @@
 
 <body>
   <noscript>
-    Please enable JavaScript to run this app.
+    You need to enable JavaScript to run this app.
   </noscript>
   <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
     crossorigin="anonymous"></script>
@@ -43,6 +59,41 @@
 
     <div id="root">
     </div>
-  </body>
+</body>
+  <!-- <style>
+  #navigator {
+    position: -webkit-sticky;
+    position: sticky;
+    top: 0;
+    z-index: 1000;
+  }
+  
+  #navigator .navbar-toggler-icon {
+    background-image: url("data:image/svg+xml;charset=utf8,%3Csvg viewBox='0 0 30 30' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath stroke='rgba(0, 0, 0, 0.5)' stroke-width='2' stroke-linecap='round' stroke-miterlimit='10' d='M4 7h22M4 15h22M4 23h22'/%3E%3C/svg%3E");
+  }
+  
+  @media (max-width: 720px) {
+    #navigator {
+      border-bottom: 1px solid #ccc;
+    }
+  }
+  
+  #navigator a {
+    font-size: 14px;
+  }
+  
+  #navigator + section {
+    padding: 250px 0;
+  }
+  </style> -->
+
+    <!-- <title>Allergenius</title> -->
+
+  <!-- <body>
+      <div className="login-bkdg">
+        <div id="root">
+        </div>
+      </div>
+</body> -->
 
 </html>
