+++ client/src/assets/css/forms.css
@@ -1,7 +1,8 @@
-/* Forms */
 form {
    margin-top: 40px;
    padding: 20px 30px 30px 20px;
+   /* border: 2px solid #3174AD;
+   background-color: #eaf6ff; */
    border-radius: 5px;
 }
 
@@ -30,9 +31,10 @@ form {
    padding: 0 auto;
 }
 
+
 .form-label {
    font-size: 19px;
-   color: #16466d;
+   color: #34495E;
 }
 
 /* Radio Buttons */
@@ -40,8 +42,15 @@ form {
    margin: 0 auto;
    width: 100%;
    float: left;
-   color: #16466d;
+   color: #203B39;
 }
+
+/* .radio-Q-label {
+   margin: 0 auto;
+   width: 100%;
+   float: left;
+   color: #203B39;
+} */
 
 .form-check-5 {
    width: 90%;
