+++ client/src/assets/css/index.css
@@ -1,6 +1,10 @@
 body {
   margin: 0;
   padding: 0;
+  /* font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+    sans-serif; */
+  /* font-family: 'Poppins', sans-serif; */
   font-family: 'Montserrat', sans-serif;
   font-family: 'Open Sans', sans-serif;
   -webkit-font-smoothing: antialiased;
