+++ client/src/components/Header/Header.js
@@ -2,7 +2,7 @@ import React from "react";
 
 const Header = props => (
     <div className="header-styling text-center">
-        <p className="text-capitalize">Welcome, {props.username}</p>
+        <p className="text-capitalize">Welcome {props.username}! </p>
     </div>
 )
