+++ client/src/index.js
@@ -4,6 +4,8 @@ import './assets/css/index.css';
 
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
+// import Routes from './auth/Routes';
+// import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
    <App />,
