+++ client/src/pages/AddProfile/AddProfile.js
@@ -32,6 +32,8 @@ class AddProfile extends Component {
 				const token = localStorage.usertoken
 				const decoded = jwt_decode(token)
 				this.setState({
+					// first_name: decoded.first_name,
+					// last_name: decoded.last_name,
 					email: decoded.email,
 					id: decoded.id
 				})
