+++ client/src/pages/EditProfile/EditProfile.js
@@ -40,6 +40,8 @@ class EditProfile extends Component {
 		const token = localStorage.usertoken
 		const decoded = jwt_decode(token)
 		this.setState({
+			// first_name: decoded.first_name,
+			// last_name: decoded.last_name,
 			email: decoded.email,
 			id: decoded.id
 		})
