+++ client/src/pages/ReactionForm/ReactionForm.css
@@ -1,3 +1,3 @@
 body {
-    background-color: #FDFFFC !important;
+    background-color: #eaf6ff !important;
 }
