+++ client/src/pages/ReactionPage/ReactionPage.css
@@ -1,7 +1,9 @@
 body {
     padding-bottom: 80px;
     margin: 0 auto;
+    /* background-image: url("../../assets/images/foodboard_1100x734.jpg"); */
     background-position: top;
-    background-color: #FDFFFC;
+    background-color: #eaf6ff;
     width: 100%;
+    /* height: 100%; */
 }
