+++ client/src/pages/ReactionPage/ReactionPage.js
@@ -35,6 +35,8 @@ class ReactionPage extends Component {
         const token = localStorage.usertoken
         const decoded = jwt_decode(token)
         this.setState({
+            // first_name: decoded.first_name,
+            // last_name: decoded.last_name,
             email: decoded.email,
             id: decoded.id
         })
