+++ client/src/pages/Register/Register.js
@@ -34,6 +34,8 @@ class Register extends Component {
         })
     }
         
+    // TODO: add check/alert for when a user already exists
+
     render () {
         return (
             <div className="container">
