+++ config/db.js
@@ -5,13 +5,15 @@ const dotenv = require('dotenv');
 
 dotenv.config()
 
+
 if (process.env.JAWSDB_URL){
     var sequelize = new Sequelize(process.env.JAWSDB_URL)
+
 } else {
     var sequelize = new Sequelize("allergenius_db", "root", "root", {
         host: 'localhost',
         dialect: 'mysql',
-        port: '8889', 
+        port: '8889', //change for Mac
         operatorsAliases: false,
 
         pool: {
@@ -22,6 +24,7 @@ if (process.env.JAWSDB_URL){
         }
     });
 }
+
 
 db.sequelize = sequelize
 db.Sequelize = Sequelize
